+++ src/unit_tests/rules/decisionminimumvertexcover_hamiltoniancircuit.rs
@@ -10,7 +10,7 @@ fn decision_mvc(
     num_vertices: usize,
     edges: &[(usize, usize)],
     weights: &[i32],
-    k: i32,
+    k: i64,
 ) -> Decision<MinimumVertexCover<SimpleGraph, i32>> {
     Decision::new(
         MinimumVertexCover::new(
